moves linode requests and adds read error to dashboard
