+++ src/App.tsx
@@ -33,7 +33,7 @@ import { requestAccountSettings } from 'src/store/accountSettings/accountSetting
 import { getAllBuckets } from 'src/store/bucket/bucket.requests';
 import { requestDomains } from 'src/store/domains/domains.actions';
 import { requestImages } from 'src/store/image/image.requests';
-import { requestLinodes } from 'src/store/linodes/linodes.actions';
+import { requestLinodes } from 'src/store/linodes/linode.requests';
 import { requestTypes } from 'src/store/linodeType/linodeType.requests';
 import { requestNotifications } from 'src/store/notification/notification.requests';
 import { requestProfile } from 'src/store/profile/profile.requests';
