+++ src/store/linodes/linode.requests.ts
@@ -1,17 +1,25 @@
+import * as Bluebird from 'bluebird';
+import requestMostRecentBackupForLinode from 'src/features/linodes/LinodesLanding/requestMostRecentBackupForLinode';
 import {
   createLinode as _createLinode,
   deleteLinode as _deleteLinode,
   getLinode as _getLinode,
+  getLinodes,
   linodeReboot as _rebootLinode,
   updateLinode as _updateLinode
 } from 'src/services/linodes';
+import { getAPIErrorOrDefault } from 'src/utilities/errorUtils';
+import { getAll } from 'src/utilities/getAll';
 import { createRequestThunk } from '../store.helpers';
+import { ThunkActionCreator } from '../types';
 import {
   createLinodeActions,
   deleteLinodeActions,
   getLinodeActions,
+  getLinodesActions,
   rebootLinodeActions,
-  updateLinodeActions
+  updateLinodeActions,
+  upsertLinode
 } from './linodes.actions';
 
 export const getLinode = createRequestThunk(getLinodeActions, ({ linodeId }) =>
@@ -36,3 +44,40 @@ export const rebootLinode = createRequestThunk(
   rebootLinodeActions,
   ({ linodeId, configId }) => _rebootLinode(linodeId, configId)
 );
+
+export const requestLinodes: ThunkActionCreator<
+  Promise<Linode.Linode[]>
+> = () => dispatch => {
+  dispatch(getLinodesActions.started);
+
+  return getAll<Linode.Linode>(getLinodes)()
+    .then(getBackupsForLinodes)
+    .then(result => {
+      dispatch(getLinodesActions.done({ result }));
+      return result;
+    })
+    .catch(err => {
+      dispatch(
+        getLinodesActions.failed({
+          error: getAPIErrorOrDefault(
+            err,
+            'There was an error retrieving your Linodes.'
+          )
+        })
+      );
+      return err;
+    });
+};
+
+const getBackupsForLinodes = ({ data }: { data: Linode.Linode[] }) =>
+  Bluebird.map(data, requestMostRecentBackupForLinode);
+
+type RequestLinodeForStoreThunk = ThunkActionCreator<void>;
+export const requestLinodeForStore: RequestLinodeForStoreThunk = id => dispatch => {
+  _getLinode(id)
+    .then(response => response.data)
+    .then(requestMostRecentBackupForLinode)
+    .then(linode => {
+      return dispatch(upsertLinode(linode));
+    });
+};
