+++ src/store/linodes/linodes.events.ts
@@ -1,12 +1,9 @@
 import { Dispatch } from 'redux';
 import { ApplicationState } from 'src/store';
+import { requestLinodeForStore } from 'src/store/linodes/linode.requests';
 import { EventHandler } from 'src/store/types';
 import { requestNotifications } from '../notification/notification.requests';
-import {
-  deleteLinode,
-  requestLinodeForStore,
-  updateLinode
-} from './linodes.actions';
+import { deleteLinode, updateLinode } from './linodes.actions';
 
 const linodeEventsHandler: EventHandler = (event, dispatch, getState) => {
   const { action, entity, status } = event;
